+++ omechu-app/src/app/restaurant/restaurant-detail/[id]/components/RestaurantDetailHeader.tsx
