+++ omechu-app/src/app/restaurant/restaurant-detail/[id]/map/components/RestaurantMapPreview.tsx
